+++ test/src/editor_test.dart
@@ -39,6 +39,7 @@ void testEditNullThrows() {
 void testDispose() {
   editor.onBlur.listen(noop1, onDone: expectAsync0(noop0));
   editor.onChange.listen(noop1, onDone: expectAsync0(noop0));
+  editor.onChangeSelection.listen(noop1, onDone: expectAsync0(noop0));
   editor.onChangeSession.listen(noop1, onDone: expectAsync0(noop0));
   editor.onCopy.listen(noop1, onDone: expectAsync0(noop0));
   editor.onFocus.listen(noop1, onDone: expectAsync0(noop0));
@@ -462,6 +463,8 @@ void testSelectAll() {
       sampleTextLines[sampleTextLines.length - 1].length);  
   expect(editor.cursorPosition, equals(const Point(0, 0)));
   expect(editor.selectionRange, equals(new Range(0, 0, 0, 0)));
+  // TODO(rms): investigate why `onChangeSelection` fires 3 times.
+  editor.onChangeSelection.listen(expectAsync1(noop1, count: 3));
   editor.selectAll();
   expect(editor.cursorPosition, equals(endCursor));
   expect(editor.selectionRange,